requirements.txt
bitrix24-rest==1.1.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.9
